+++ meuMenu/src/main/java/meumenu/application/meumenu/controllers/UsuarioController.java
@@ -2,9 +2,11 @@ package meumenu.application.meumenu.controllers;
 
 import jakarta.validation.Valid;
 import meumenu.application.meumenu.restaurante.Restaurante;
+import meumenu.application.meumenu.restaurante.RestauranteDTO;
 import meumenu.application.meumenu.restaurante.RestauranteRepository;
 import meumenu.application.meumenu.usuario.UsuarioDTO;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import meumenu.application.meumenu.usuario.DadosCadastroUsuario;
@@ -25,71 +27,95 @@ public class UsuarioController {
 
     @PostMapping("/cadastrar")
     @Transactional
-    public void cadastrar(@RequestBody @Valid DadosCadastroUsuario dados){
+    public ResponseEntity<UsuarioDTO> cadastrar(@RequestBody @Valid DadosCadastroUsuario dados){
         repository.save(new Usuario(dados));
+        List<Usuario> tempUsuario = repository.findAll();
+        UsuarioDTO usuario = new UsuarioDTO(tempUsuario.get(tempUsuario.size()-1).getId(),dados.nome(), dados.sobrenome(), dados.email(), dados.tipoComidaPreferida().name());
+        return ResponseEntity.status(200).body(usuario);
     }
 
     @GetMapping
     @Transactional
-    public List<UsuarioDTO> listar(){
+    public ResponseEntity<List<UsuarioDTO>> listar(){
         List<UsuarioDTO> usuariosDTO = new ArrayList<>();
         List<Usuario> tempUsuario = repository.findAll();
         for(Usuario u : tempUsuario){
-            usuariosDTO.add(new UsuarioDTO(u.getNome(), u.getSobrenome(), u.getEmail(), u.getTipoComidaPreferida().name()));
+            usuariosDTO.add(new UsuarioDTO(u.getId(),u.getNome(), u.getSobrenome(), u.getEmail(), u.getTipoComidaPreferida().name()));
         }
-        return usuariosDTO;
+        return ResponseEntity.status(200).body(usuariosDTO);
+    }
+
+    @GetMapping("{id}")
+    @Transactional
+    public ResponseEntity<UsuarioDTO> listarPorId(@PathVariable Integer id){
+        List<Usuario> tempUsuario = repository.findAll();
+        for(Usuario u : tempUsuario){
+            if(u.getId().equals(id)){
+                UsuarioDTO usuario = new UsuarioDTO(u.getId(), u.getNome(), u.getSobrenome(), u.getEmail(), u.getTipoComidaPreferida().name());
+                return ResponseEntity.status(200).body(usuario);
+            }
+        }
+        return ResponseEntity.status(404).build();
     }
 
     @PutMapping("/{id}")
     @Transactional
-    public String atualizar(@RequestBody @Valid Usuario dados, @PathVariable int id){
-        Usuario usuario = repository.findById(id).orElseThrow();
-        usuario.setTipoComidaPreferida(dados.getTipoComidaPreferida());
-        repository.save(usuario);
-        return "Usuário atualizado com sucesso";
+    public ResponseEntity<String> atualizar(@RequestBody @Valid Usuario dados, @PathVariable int id){
+        if(repository.existsById(id)){
+            Usuario usuario = repository.findById(id).orElseThrow();
+            usuario.setTipoComidaPreferida(dados.getTipoComidaPreferida());
+            repository.save(usuario);
+            return ResponseEntity.status(200).body("Usuário atualizado com sucesso");
+        }
+        return ResponseEntity.status(404).body("Usuário não encontrado");
     }
 
     @PostMapping("/logar")
     @Transactional
-    public Usuario logar(@RequestBody Usuario dados  ){
+    public ResponseEntity<String> logar(@RequestBody Usuario dados  ){
 
         List<Usuario> listaUsuario = repository.findAll();
 
         for(Usuario b : listaUsuario){
             if(b.getEmail().equals(dados.getEmail()) && b.getSenha().equals(dados.getSenha())){
-                return b;
+                return ResponseEntity.status(200).body("Login efetuado com sucesso");
             }
         }
-        return null;
+        return ResponseEntity.status(401).body("Email e senha incorretos");
     }
 
     @DeleteMapping("/{id}")
     @Transactional
-    public String deletar(@PathVariable int id){
-        Usuario usuario = repository.findById(id).orElseThrow();
-        repository.delete(usuario);
-
-        return "Usuário deletado com sucesso";
+    public ResponseEntity<String> deletar(@PathVariable int id){
+        if(repository.existsById(id)){
+            Usuario usuario = repository.findById(id).orElseThrow();
+            repository.delete(usuario);
+            return ResponseEntity.status(200).body("Usuário deletado com sucesso");
+        }
+        return ResponseEntity.status(404).body("Usuário não encontrado");
     }
 
     @GetMapping("/recomendar/{id}")
     @Transactional
-    public List<Usuario> recomendar(@PathVariable int id){
+    public ResponseEntity<List<RestauranteDTO>> recomendar(@PathVariable int id){
         List<UsuarioDTO> usuariosDTO = new ArrayList<>();
-
         List<Usuario> listaUsuario = repository.findAll();
         List<Restaurante> listaRestaurante = repositoryRestaurante.findAll();
+        List<RestauranteDTO> restauranteDTO = new ArrayList<>();
 
-        for(Usuario b : listaUsuario){
-            if(b.getId() == id){
-                return b.recomendar(listaUsuario, listaRestaurante, id);
+        if(!repository.existsById(id)){
+            return ResponseEntity.status(404).build();
+        }
+
+        for(Usuario usuario : listaUsuario){
+            if(usuario.getId() == id){
+                restauranteDTO = (usuario.recomendar(listaUsuario, listaRestaurante, id));
             }
         }
-        return null;
+        if(restauranteDTO.isEmpty()){
+            return ResponseEntity.status(204).build();
+        }
+            return ResponseEntity.status(200).body(restauranteDTO);
     }
-
-
-
-
 
 }
